fix: corrected anti-collision lights flashing behavior in player aircraft
anti-collision light was flashing due to a conflict in behaviour between AI and player model. New animation args were added to the external model to deconflict the AI and player anti-collision lights. Due to lack of compatible light type for AI model, AI anti-collision light will be a red strobe instead of a rotary beacon.

Closes #175