refactor(input): import common bindings from base sim files and removed duplicate bindings from input files
Keyboard and joystick bindings that can be found in common_joystick_bindings.lua and common_keyboard_binding.lua are removed from the input files and imported from the base sim files instead. This allows for a cleaner set of input binding files and consistency with the base sim commands. This would also prevent duplicate commands.