Fixed Brake issue
My (JNelson) brakes on my pedals were mapped from 0->1.0 however DCS expects -1.0 -> 1.0 mapping. It turns out there is a bug or something in the pedal software that causes inverting the brakes to have this effect of mapping 0 -> 1.0. Removing this option and the problem is fixed.

The brakes have now been correctly mapped from -1.0 -> 1.0. The
direction was preserved so people updating shouldn't have to invert.