Fixed NWS range
Issue spotted in #304.

The nosewheel does not align correctly. This is because the EFM
suspension position and animation were not lining up. This was for
several reasons.

The angle range was unclear thinking it was -90 to +90 the rudder input
was divided by 4. This actually resulted in a NWS angle which was only
22.5 degrees. It seems the animation is completely unrelated to the
physical simulation. As such the correct factor is 2.

The animation can be aligned by changing the tand_gear_max a value which
is completely separate from the EFM suspension code.