Moved EFM commands
EFM commands have been moved to the start of the Keys table definition.

This is so that commands can be added after this without changing the
value of these.

These must be built directly into the dll because runtime values cannot
be used with switch statements. The Generate_Control_Headers.exe will
generate the Commands.h and Devices.h headers containing the key and
device commands and the device numbers.