Initial ICLS and Tacan
This commit implements mobile tacan and icls. It still has a long way to
go.

This enables Tacan for tankers and carriers without workarounds. One
limitation of this approach is that it relies on the mission data. So
any spawned objects must have the same callsign as a similar unit in the
mission to be detected.