cockpit file rename bonanza
assets are uploading to google drive now. this is the first step in resolving issue #312 -- heclak and i are keeping track on a google sheet of the process, but we'll have one more go-round where a few files will change.
several old and unneeded files were deprecated and we are far more conflict-proof in the future.