Fix issue #392 MCL course-needle offset on HMAS Melbourne
Propagate the ship type type from the mission description to the MCL handling.

There try to find the ship type in a table and get it's beacon offsets and flight deck angle.  If not found, use the Stennis data as default values.

Deck angle for HMAS Melbourne from  https://www.navy.gov.au/history/angled-flight-deck but seems a good match. The offsets are found experimentally and could be inaccurate.