Overhall CG Modelling
Before CG was fixed at 0,0,0 which lead to no moment produced by the
wing as its CP was also at 0,0,0.

The elevator was being stalled since its throw was set to 30 degrees
(instead of 17 degrees). Elevator was separated out into standard moment
so it can be tuned independently of the stab.

Tanks are now in the correct positions to move the CG.
Amount CG moves when drained:
- Fuslage tank -> 7% aft
- Wing tank -> 6% fwd
- External Tank -> 1 % fwd