ADF and Guard Radio
Guard Radio is additional radio (receiver) in the plane. Slightly ficticious guard transmit option added to make it easier and prevent users having to switch to GRD XMIT when wanting to transmit on guard.

ADF transmits a direction finding tone so nothing will be heard by others when on ADF however any reception will be garbled (TODO).

Move setup of radios to intercom so it's only completed once.